Time: 7 ms (83.69%), Space: 10.2 MB (95.23%) - LeetHub

+++ 0518-coin-change-ii/0518-coin-change-ii.cpp
@@ -2,18 +2,17 @@ class Solution {
 public:
     int change(int amount, vector<int>& coins) {
         int n = coins.size();
-        vector<long long> dp(amount + 1, 0);
+        vector<int> dp(amount + 1, 0);
 
         dp[0] = 1;
 
         for (int i = 0; i < n; i++) {
             for (int c = coins[i]; c <= amount; c++) {
-                if (dp[c] <= LLONG_MAX - dp[c - coins[i]]) {
+                if (dp[c] <= INT_MAX - dp[c - coins[i]]) {
                     dp[c] += dp[c - coins[i]];
                 }
             }
         }
-
-        return (int)dp[amount];
+        return dp[amount];
     }
 };
